+++ src/org/fxyz/shapes/primitives/SpringMesh.java
@@ -396,10 +396,10 @@ public class SpringMesh extends TexturedMesh {
         return createMesh();
     }
  
-    public float getTau(double t){
+    public double getTau(double t){
         return spring.getTau(t);
     }
-    public float getKappa(double t){
+    public double getKappa(double t){
         return spring.getKappa(t);
     }
